+++ crates/nerd-font-symbols/src/logos.rs
@@ -6,18 +6,36 @@ pub const LINUX_ALPINE: &str = "";
 pub const LINUX_AOSC: &str = "";
 /// ``
 pub const LINUX_APPLE: &str = "";
+/// ``
+pub const LINUX_ARCHCRAFT: &str = "";
 /// ``
 pub const LINUX_ARCHLABS: &str = "";
 /// ``
 pub const LINUX_ARCHLINUX: &str = "";
+/// ``
+pub const LINUX_ARCOLINUX: &str = "";
+/// ``
+pub const LINUX_ARDUINO: &str = "";
 /// ``
 pub const LINUX_ARTIX: &str = "";
+/// ``
+pub const LINUX_AWESOME: &str = "";
+/// ``
+pub const LINUX_BIGLINUX: &str = "";
+/// ``
+pub const LINUX_BSPWM: &str = "";
 /// ``
 pub const LINUX_BUDGIE: &str = "";
 /// ``
 pub const LINUX_CENTOS: &str = "";
+/// ``
+pub const LINUX_CINNAMON: &str = "";
+/// ``
+pub const LINUX_CODEBERG: &str = "";
 /// ``
 pub const LINUX_COREOS: &str = "";
+/// ``
+pub const LINUX_CRYSTAL: &str = "";
 /// ``
 pub const LINUX_DEBIAN: &str = "";
 /// ``
@@ -26,10 +44,16 @@ pub const LINUX_DEEPIN: &str = "";
 pub const LINUX_DEVUAN: &str = "";
 /// ``
 pub const LINUX_DOCKER: &str = "";
+/// ``
+pub const LINUX_DWM: &str = "";
 /// ``
 pub const LINUX_ELEMENTARY: &str = "";
 /// ``
 pub const LINUX_ENDEAVOUR: &str = "";
+/// ``
+pub const LINUX_ENLIGHTENMENT: &str = "";
+/// ``
+pub const LINUX_FDROID: &str = "";
 /// ``
 pub const LINUX_FEDORA: &str = "";
 /// ``
@@ -38,40 +62,130 @@ pub const LINUX_FEDORA_INVERSE: &str = "";
 pub const LINUX_FERRIS: &str = "";
 /// ``
 pub const LINUX_FLATHUB: &str = "";
+/// ``
+pub const LINUX_FLUXBOX: &str = "";
+/// ``
+pub const LINUX_FORGEJO: &str = "";
+/// ``
+pub const LINUX_FOSDEM: &str = "";
 /// ``
 pub const LINUX_FREEBSD: &str = "";
+/// ``
+pub const LINUX_FREECAD: &str = "";
+/// ``
+pub const LINUX_FREEDESKTOP: &str = "";
+/// ``
+pub const LINUX_GARUDA: &str = "";
 /// ``
 pub const LINUX_GENTOO: &str = "";
+/// ``
+pub const LINUX_GIMP: &str = "";
+/// ``
+pub const LINUX_GITEA: &str = "";
+/// ``
+pub const LINUX_GNOME: &str = "";
 /// ``
 pub const LINUX_GNU_GUIX: &str = "";
+/// ``
+pub const LINUX_GTK: &str = "";
+/// ``
+pub const LINUX_HYPERBOLA: &str = "";
+/// ``
+pub const LINUX_HYPRLAND: &str = "";
+/// ``
+pub const LINUX_I3: &str = "";
 /// ``
 pub const LINUX_ILLUMOS: &str = "";
+/// ``
+pub const LINUX_INKSCAPE: &str = "";
+/// ``
+pub const LINUX_JWM: &str = "";
 /// ``
 pub const LINUX_KALI_LINUX: &str = "";
+/// ``
+pub const LINUX_KDE: &str = "";
+/// ``
+pub const LINUX_KDE_NEON: &str = "";
+/// ``
+pub const LINUX_KDE_PLASMA: &str = "";
+/// ``
+pub const LINUX_KDENLIVE: &str = "";
+/// ``
+pub const LINUX_KICAD: &str = "";
+/// ``
+pub const LINUX_KRITA: &str = "";
+/// ``
+pub const LINUX_KUBUNTU: &str = "";
+/// ``
+pub const LINUX_KUBUNTU_INVERSE: &str = "";
 /// ``
 pub const LINUX_LINUXMINT: &str = "";
 /// ``
 pub const LINUX_LINUXMINT_INVERSE: &str = "";
+/// ``
+pub const LINUX_LOCOS: &str = "";
+/// ``
+pub const LINUX_LXDE: &str = "";
+/// ``
+pub const LINUX_LXLE: &str = "";
+/// ``
+pub const LINUX_LXQT: &str = "";
 /// ``
 pub const LINUX_MAGEIA: &str = "";
 /// ``
 pub const LINUX_MANDRIVA: &str = "";
 /// ``
 pub const LINUX_MANJARO: &str = "";
+/// ``
+pub const LINUX_MATE: &str = "";
+/// ``
+pub const LINUX_MPV: &str = "";
+/// ``
+pub const LINUX_MXLINUX: &str = "";
+/// ``
+pub const LINUX_NEOVIM: &str = "";
 /// ``
 pub const LINUX_NIXOS: &str = "";
+/// ``
+pub const LINUX_OCTOPRINT: &str = "";
 /// ``
 pub const LINUX_OPENBSD: &str = "";
+/// ``
+pub const LINUX_OPENSCAD: &str = "";
 /// ``
 pub const LINUX_OPENSUSE: &str = "";
+/// ``
+pub const LINUX_OSH: &str = "";
+/// ``
+pub const LINUX_OSHWA: &str = "";
+/// ``
+pub const LINUX_OSI: &str = "";
+/// ``
+pub const LINUX_PARABOLA: &str = "";
 /// ``
 pub const LINUX_PARROT: &str = "";
 /// ``
 pub const LINUX_POP_OS: &str = "";
+/// ``
+pub const LINUX_POSTMARKETOS: &str = "";
+/// ``
+pub const LINUX_PRUSASLICER: &str = "";
+/// ``
+pub const LINUX_PUPPY: &str = "";
+/// ``
+pub const LINUX_QT: &str = "";
+/// ``
+pub const LINUX_QTILE: &str = "";
+/// ``
+pub const LINUX_QUBESOS: &str = "";
 /// ``
 pub const LINUX_RASPBERRY_PI: &str = "";
 /// ``
 pub const LINUX_REDHAT: &str = "";
+/// ``
+pub const LINUX_REPRAP: &str = "";
+/// ``
+pub const LINUX_RISCV: &str = "";
 /// ``
 pub const LINUX_ROCKY_LINUX: &str = "";
 /// ``
@@ -84,13 +198,39 @@ pub const LINUX_SLACKWARE_INVERSE: &str = "";
 pub const LINUX_SNAPPY: &str = "";
 /// ``
 pub const LINUX_SOLUS: &str = "";
+/// ``
+pub const LINUX_SWAY: &str = "";
+/// ``
+pub const LINUX_TAILS: &str = "";
+/// ``
+pub const LINUX_THUNDERBIRD: &str = "";
+/// ``
+pub const LINUX_TOR: &str = "";
+/// ``
+pub const LINUX_TRISQUEL: &str = "";
 /// ``
 pub const LINUX_TUX: &str = "";
 /// ``
 pub const LINUX_UBUNTU: &str = "";
 /// ``
 pub const LINUX_UBUNTU_INVERSE: &str = "";
+/// ``
+pub const LINUX_VANILLA: &str = "";
 /// ``
 pub const LINUX_VOID: &str = "";
+/// ``
+pub const LINUX_VSCODIUM: &str = "";
+/// ``
+pub const LINUX_WAYLAND: &str = "";
+/// ``
+pub const LINUX_WIKIMEDIA: &str = "";
+/// ``
+pub const LINUX_XEROLINUX: &str = "";
+/// ``
+pub const LINUX_XFCE: &str = "";
+/// ``
+pub const LINUX_XMONAD: &str = "";
+/// ``
+pub const LINUX_XORG: &str = "";
 /// ``
 pub const LINUX_ZORIN: &str = "";